+++ packages/camera/camera/lib/camera.dart
@@ -11,5 +11,7 @@ export 'package:camera_platform_interface/camera_platform_interface.dart'
         CameraDescription,
         CameraException,
         CameraLensDirection,
+        FlashMode,
+        ExposureMode,
         ResolutionPreset,
         XFile;